Make calls to postgresql's pg_terminate_backed() an --fbt option, rather than the default
During development, this was required because intermine wasn't closing its connections properly.
This may have been resolved by intermine commit 9d361b90, where the postprocess is now being proactive in closing the connection pool rather than hooking into jvm exit.
However, we will keep this option as it may be useful if the problem reoccurs

+++ interminepy/utils.py
@@ -108,12 +108,13 @@ def pg_terminate_backends(db_configs, options):
 
 
 def pg_terminate_backend(db_config, options):
-    run_on_db(
-        ['psql',
-         '-P', 'pager=off',
-         '-q',
-         '-o', '/dev/null',
-         '-c', 'SELECT pg_terminate_backend(pg_stat_activity.pid)'
-               ' FROM pg_stat_activity '
-               ' WHERE datname = current_database()'
-               ' AND pid <> pg_backend_pid();', db_config['name']], db_config, options)
+    if (options['force-backend-termination']):
+        run_on_db(
+            ['psql',
+             '-P', 'pager=off',
+             '-q',
+             '-o', '/dev/null',
+             '-c', 'SELECT pg_terminate_backend(pg_stat_activity.pid)'
+                   ' FROM pg_stat_activity '
+                   ' WHERE datname = current_database()'
+                   ' AND pid <> pg_backend_pid();', db_config['name']], db_config, options)
